Kubernetes: Move over tests
The kubernetes module that replaces the deprecated client methods only
has a test for login functionality. Since the test coverage was broader
for the client methods, the client method tests have been moved over to
test_kubernetes to provide extra coverage. To remove duplicated code,
the client kubernetes tests have been deleted.
